refactor(ai-service): Step 6 — rehome voice modality to app/voice/
The Pipecat / Kokoro / STT-TTS surface moves out of services/voice/
into its own top-level band, matching the diagram (Page 3).

Moves (whole subpackage):
  app/services/voice/__init__.py        →  app/voice/__init__.py
  app/services/voice/kokoro_service.py  →  app/voice/kokoro_service.py
  app/services/voice/pipecat_kokoro.py  →  app/voice/pipecat_kokoro.py
  app/services/voice/pipecat_pipeline.py→  app/voice/pipecat_pipeline.py
  app/services/voice/sentence_streamer.py → app/voice/sentence_streamer.py
  app/services/voice/stream_events.py   →  app/voice/stream_events.py

Empty placeholder subdirs app/voice/stt/ and app/voice/tts/ removed
for now; faster-whisper STT currently lives inside pipecat_pipeline,
and Kokoro TTS lives in kokoro_service.py. Those split into stt/ and
tts/ subpackages later if the surface grows.

Importers updated (6 files):
  • app/main not affected (voice is imported via api/voice.py)
  • app/api/{voice,query}.py
  • app/services/ai_service.py  (voice streaming integration)
  • app/voice/{__init__, pipecat_pipeline, pipecat_kokoro}.py  (internal absolute refs)

Validated: /health 200, /voice/health 200, app.main + voice imports clean.
