+++ acla_ai_service/app/services/ai_service.py
@@ -11,9 +11,9 @@ from openai import AsyncOpenAI
 from app.services.full_dataset_ml_service import Full_dataset_TelemetryMLService
 from app.core import settings
 from app.integrations.backend.client import BackendService
-from app.services.voice import get_kokoro_service
-from app.services.voice.sentence_streamer import SentenceStreamer
-from app.services.voice.stream_events import (
+from app.voice import get_kokoro_service
+from app.voice.sentence_streamer import SentenceStreamer
+from app.voice.stream_events import (
     event_audio,
     event_done,
     event_error,
