--- acla_ai_service/app/services/voice/__init__.py
@@ -1,5 +1,0 @@
-"""Voice synthesis and conversation services (Phase 2+)."""
-
-from app.services.voice.kokoro_service import KokoroService, get_kokoro_service
-
-__all__ = ["KokoroService", "get_kokoro_service"]
